net:wireless:mwifiex: Make AP interface optional
By default both station interface (mlan) and host ap interface (uap) are
created by the driver. This patch makes it optional to select whether to
enable ap interface or not.

The reason for doing this is, both mlan and uap interface use the same
mac address and connman network manager mistakenly interpret connection
profile of mlan interface to be of uap and gives error. Which intern
causes wireless connection failure.

Signed-off-by: Ajay Bhargav <bhargav.ajay@gmail.com>