arm:dove-d2plug: Add GPIO LED support
There are three LEDs on D2plug:
1. WiFi AP mode - LED0
2. Wifi Station - LED1
3. Bluetooth - LED2

Though none of them are actually connected to wifi-bt module so LED
trigger was best option to utilize them.

LED0 is configured for Wifi-rx as AP mode is not used
LED1 is configured for Wifi-tx
LED2 is configured for MMC0 activity as BT is not yet functional.

Signed-off-by: Ajay Bhargav <bhargav.ajay@gmail.com>